Add view model for menu list
The view model for the menu row uses a closure (or function) for
performing the grouping of menu items. The closure is a function that
takes an array of menu items and returns and array of menu sections.
This is what the previously written groupMenuByCategory does.

To test the implementation, a spy closure is created in the test which
just sets a flag to say it was called, and returns a pre-populated
array of menu sections.

This test ensures the initilisation of the view model uses the spy
closure, which sets the view model sections property to the
inputSections array of menu sections.