Implement fetching menu using placeholder
To begin implementing publishing a new menu when fetching it from
somewhere, a placeholer is used to implement a method for fetching the
menu. This uses a protocol to implement the Dependency Inversion
Principal - providing an abstraction of the data source (which will
likely involve a network call).

In the view model, we call fetchMenu() method to fetch the data via the
abstraction (which is defined using the MenuFetching protocol). On
success, the menuGrouping closure is called to transform the menu items
into menu sections.

Since the implementation of the fetchMenu() in the placeholder uses a
publisher with a delay, the test uses an asynchronous test - which is
appropriate for a publisher.