Replace menu with a test stub version
Instead of using the MenuFetchingPlaceholder to generate an array of
menu items (which just returns a publisher containing the menu global
variable), use a stub. Here the stub just returns the array of menu
items passed when it is initialised - in this case just a single
elemenet created from the MenuItem.fixture().

This shows the power of using a protocol for the menuFetching parameter
of the view model - we can simply replace the MenuFetchingPlaceholder
with a MenuFetchingStub. The MenuFetchingStub can be created (and
modified) to do whatever we require it to do for testing without
affecting anything in the implementation.