Use a Result<> type to make errors explicit
Using a Result<[MenuSection], Error> makes it explicit the sections
property of the view model can either contiain an array of menu
sections (if the menu fetching succeeds) or an error (if the menu
fetching fails).

We also update the View to handl both cases by displaying an error
message on failure or the list of sections on success.

To cater for this change in type, we had to update the tests. This uses
two different methods of accessing the menu item array on success:
 - Using the Result type's get() method (which can throw an error).
 - Using a guard statement.