Handle error in view model when fetching
This test uses a MenuFetchingStub which is created with an error
Result instead of a success result.

Note: This requires no update to the MenuFetchingStub as it just sets
its result to the item passed when it is initialised.

The only change is to the view model to handle the error condition in
the receiveCompletion closure of the sink() method.